Add a couple of brackets for VMA
